requirements.txt
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorclass==2.2.2
defusedxml==0.7.1
docopt==0.6.2
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pip-upgrader==1.4.15
pipreqs==0.4.13
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rsa==4.9
six==1.17.0
terminaltables==3.1.10
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
yarg==0.1.10
youtube-transcript-api==0.6.3
